+++ luxonis_ml/data/utils/__init__.py
@@ -5,13 +5,14 @@ from .data_utils import (
     warn_on_duplicates,
 )
 from .enums import BucketStorage, BucketType, ImageType, MediaType
-from .label_utils import (
+from .parquet import ParquetDetection, ParquetFileManager, ParquetRecord
+from .task_utils import (
+    get_qualified_task_name,
     get_task_name,
     get_task_type,
     split_task,
     task_type_iterator,
 )
-from .parquet import ParquetDetection, ParquetFileManager, ParquetRecord
 from .types import Labels, LuxonisLoaderOutput
 from .visualizations import concat_images, create_text_image, visualize
 
@@ -21,6 +22,7 @@ __all__ = [
     "create_text_image",
     "concat_images",
     "visualize",
+    "get_qualified_task_name",
     "check_array",
     "infer_task",
     "warn_on_duplicates",
